Services now return Entity and User business rule implemented

+++ src/main/java/com/posting/post/services/UserService.java
@@ -1,17 +1,16 @@
 package com.posting.post.services;
 
-import java.util.List;
 import java.util.Optional;
 
-import com.posting.post.dto.common.PageResponseDTO;
 import com.posting.post.dto.request.UserRequestDTO;
+import com.posting.post.services.exceptions.ConflictException;
+import com.posting.post.services.exceptions.UnauthorizedActionException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
-import com.posting.post.dto.response.UserResponseDTO;
-import com.posting.post.entities.AdressUser;
 import com.posting.post.entities.User;
 import com.posting.post.mapper.UserMapper;
-import com.posting.post.repositories.AdressUserRepository;
 import com.posting.post.repositories.UserRepository;
 import com.posting.post.services.exceptions.ResourceNotFoundException;
 
@@ -24,13 +23,10 @@ public class UserService {
     UserRepository userRepository;
 
     @Autowired
-    AdressUserRepository adressUserRepository;
-
-    @Autowired
     UserMapper userMapper;
 
-    public List<User> findAll() {
-        return userRepository.findAll();
+    public Page<User> findAll(int page, int size) {
+        return userRepository.findAll(PageRequest.of(page, size));
     }
 
     public User findById(Long id) {
@@ -38,24 +34,42 @@ public class UserService {
         return user.orElseThrow(() -> new ResourceNotFoundException(user));
     }
 
-    public UserResponseDTO insert(UserRequestDTO dto) {
+    public User insert(UserRequestDTO dto) {
+
+        // Futura Refatoração com Código por Email
         User user = userMapper.toEntity(dto);
-        userRepository.save(user);
-        return userMapper.toDto(user);
+        boolean exists = userRepository.existsByEmail(user.getEmail());
+        if (exists) {
+            throw new ConflictException(user.getEmail(), "Email já cadastrado!");
+        }
+
+        return userRepository.save(user);
     }
 
     public User update(Long id, User obj) {
+        // Regra de negócio
+        boolean exists = userRepository.existsById(id);
+        if (!exists) {
+            throw new UnauthorizedActionException(id);
+        }
+
         User entity = userRepository.getReferenceById(id);
         updateData(entity, obj);
         return userRepository.save(entity);
     }
 
-    public void delete(Long id) {
-        userRepository.deleteById(id);
-    }
-
+    // Auxílio Update
     public void updateData(User entity, User obj) {
         entity.setName(obj.getName());
         entity.setEmail(obj.getEmail());
+    }
+
+    public void delete(Long id) {
+        User user = userRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException(id));
+
+        if (!user.getId().equals(id)) {
+            throw new UnauthorizedActionException(id);
+        }
+        userRepository.deleteById(id);
     }
 }
